apps/backend/requirements.txt

# Core
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt==4.0.1
google-generativeai>=0.3.0
pypdf>=3.17.0
python-docx>=1.1.0  # Pin version compatible with passlib if possible, but pbkdf2 is safer
pinecone-client>=3.0.0

# External Services
httpx>=0.25.0
aiofiles>=23.0.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
redis>=5.0.0
hiredis>=2.3.0
python-jobspy>=1.1.51
motor>=3.3.2

# Job Scheduler
apscheduler>=3.10.0

# Integrations (Optional)
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.26
aiosmtplib>=3.0.1

# ── Engine de Vagas / AI Analyzer / MCP Server ──────────────────────────────
openai>=1.0.0          # Análise de vagas com GPT-4o-mini (job_ai_analyzer.py)
mcp>=1.0.0             # Servidor MCP para Claude Desktop (opcional, instalar se usar mcp_server_vagas.py)
